feat: redesign layout with vertical toolbar
- Replace toolbar and 2 collumn layout  with vertical 3-column layout
- Add vertical toolbar with sections: working modes, marking types, tools
- Move rotation to working modes section
