+++ src/lib/locales/en/cursor.ts
@@ -4,6 +4,7 @@ const d: Dictionary = {
     Mode: {
         Marking: "Marking mode",
         Selection: "Selection mode",
+        Rotation: "Rotation mode",
     },
 };
