+++ src/lib/locales/en/keywords.ts
@@ -10,6 +10,7 @@ const d: Dictionary = {
     Types: "Types",
     Remove: "Remove",
     Add: "Add",
+    Tools: "Tools",
 };
 
 export default d;
